chore: split out ContactListItem and make it testable
